Products with modals

+++ Shopping/Shopping/Controllers/ProductsController.cs
@@ -6,6 +6,7 @@ using Shopping.Data.Entities;
 using Shopping.Helpers;
 using Shopping.Models;
 using Vereyon.Web;
+using static Shopping.Helpers.ModalHelper;
 
 namespace Shopping.Controllers
 {
@@ -35,6 +36,8 @@ namespace Shopping.Controllers
                 .ToListAsync());
         }
 
+
+        [NoDirectAccess]
         public async Task<IActionResult> Create()
         {
             CreateProductViewModel model = new()
@@ -83,7 +86,16 @@ namespace Shopping.Controllers
                 {
                     _ = _context.Add(product);
                     _ = await _context.SaveChangesAsync();
-                    return RedirectToAction(nameof(Index));
+                    _flashMessage.Confirmation("Registro creado.");
+                    return Json(new
+                    {
+                        isValid = true,
+                        html = ModalHelper.RenderRazorViewToString(this, "_ViewAllProducts", _context.Products
+                        .Include(p => p.ProductImages)
+                        .Include(p => p.ProductCategories)
+                        .ThenInclude(pc => pc.Category).ToList())
+                    });
+
                 }
                 catch (DbUpdateException dbUpdateException)
                 {
@@ -102,15 +114,12 @@ namespace Shopping.Controllers
                 }
             }
 
-            model.Categories = await _combosHelper.GetComboCategoriesAsync();
-            return View(model);
+            return Json(new { isValid = false, html = ModalHelper.RenderRazorViewToString(this, "Create", model) });
         }
-        public async Task<IActionResult> Edit(int? id)
+        [NoDirectAccess]
+        public async Task<IActionResult> Edit(int id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+
 
             Product product = await _context.Products.FindAsync(id);
             if (product == null)
@@ -148,7 +157,16 @@ namespace Shopping.Controllers
                 product.Stock = model.Stock;
                 _ = _context.Update(product);
                 _ = await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                _flashMessage.Confirmation("Registro actualizado.");
+                return Json(new
+                {
+                    isValid = true,
+                    html = ModalHelper.RenderRazorViewToString(this, "_ViewAllProducts", _context.Products
+                    .Include(p => p.ProductImages)
+                    .Include(p => p.ProductCategories)
+                    .ThenInclude(pc => pc.Category).ToList())
+                });
+
             }
             catch (DbUpdateException dbUpdateException)
             {
@@ -166,7 +184,7 @@ namespace Shopping.Controllers
                 _flashMessage.Danger(string.Empty, exception.Message);
             }
 
-            return View(model);
+            return Json(new { isValid = false, html = ModalHelper.RenderRazorViewToString(this, "Edit", model) });
         }
         public async Task<IActionResult> Details(int? id)
         {
@@ -182,12 +200,9 @@ namespace Shopping.Controllers
                 .FirstOrDefaultAsync(p => p.Id == id);
             return product == null ? NotFound() : View(product);
         }
-        public async Task<IActionResult> AddImage(int? id)
+        [NoDirectAccess]
+        public async Task<IActionResult> AddImage(int id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
 
             Product product = await _context.Products.FindAsync(id);
             if (product == null)
@@ -221,7 +236,17 @@ namespace Shopping.Controllers
                 {
                     _ = _context.Add(productImage);
                     _ = await _context.SaveChangesAsync();
-                    return RedirectToAction(nameof(Details), new { product.Id });
+                    _flashMessage.Confirmation("Imagen Agregada.");
+                    return Json(new
+                    {
+                        isValid = true,
+                        html = ModalHelper.RenderRazorViewToString(this, "Details", _context.Products
+                            .Include(p => p.ProductImages)
+                            .Include(p => p.ProductCategories)
+                            .ThenInclude(pc => pc.Category)
+                            .FirstOrDefaultAsync(p => p.Id == model.ProductId))
+                    });
+
                 }
                 catch (Exception exception)
                 {
@@ -229,7 +254,7 @@ namespace Shopping.Controllers
                 }
             }
 
-            return View(model);
+            return Json(new { isValid = false, html = ModalHelper.RenderRazorViewToString(this, "AddImage", model) });
         }
         public async Task<IActionResult> DeleteImage(int? id)
         {
@@ -249,8 +274,11 @@ namespace Shopping.Controllers
             await _blobHelper.DeleteBlobAsync(productImage.ImageId, "products");
             _ = _context.ProductImages.Remove(productImage);
             _ = await _context.SaveChangesAsync();
+            _flashMessage.Info("Registro borrado.");
             return RedirectToAction(nameof(Details), new { productImage.Product.Id });
         }
+
+        [NoDirectAccess]
         public async Task<IActionResult> AddCategory(int? id)
         {
             if (id == null)
@@ -304,7 +332,17 @@ namespace Shopping.Controllers
                 {
                     _ = _context.Add(productCategory);
                     _ = await _context.SaveChangesAsync();
-                    return RedirectToAction(nameof(Details), new { product.Id });
+                    _flashMessage.Confirmation("Categoria agregada.");
+                    return Json(new
+                    {
+                        isValid = true,
+                        html = ModalHelper.RenderRazorViewToString(this, "Details", _context.Products
+                            .Include(p => p.ProductImages)
+                            .Include(p => p.ProductCategories)
+                            .ThenInclude(pc => pc.Category)
+                            .FirstOrDefaultAsync(p => p.Id == model.ProductId))
+                    });
+
                 }
                 catch (Exception exception)
                 {
@@ -321,7 +359,7 @@ namespace Shopping.Controllers
             }).ToList();
 
             model.Categories = await _combosHelper.GetComboCategoriesAsync(categories);
-            return View(model);
+            return Json(new { isValid = false, html = ModalHelper.RenderRazorViewToString(this, "AddCategory", model) });
         }
 
         public async Task<IActionResult> DeleteCategory(int? id)
@@ -341,45 +379,34 @@ namespace Shopping.Controllers
 
             _ = _context.ProductCategories.Remove(productCategory);
             _ = await _context.SaveChangesAsync();
+            _flashMessage.Info("Registro borrado.");
             return RedirectToAction(nameof(Details), new { productCategory.Product.Id });
         }
 
 
-        public async Task<IActionResult> Delete(int? id)
+        [NoDirectAccess]
+        public async Task<IActionResult> Delete(int id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
             Product product = await _context.Products
                 .Include(p => p.ProductCategories)
                 .Include(p => p.ProductImages)
                 .FirstOrDefaultAsync(p => p.Id == id);
-            return product == null ? NotFound() : View(product);
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Delete(Product model)
-        {
-            Product product = await _context.Products
-                .Include(p => p.ProductImages)
-                .Include(p => p.ProductCategories)
-                .FirstOrDefaultAsync(p => p.Id == model.Id);
-
-
-            _ = _context.Products.Remove(product);
-            _ = await _context.SaveChangesAsync();
-
+            if (product == null)
+            {
+                return NotFound();
+            }
 
             foreach (ProductImage productImage in product.ProductImages)
             {
                 await _blobHelper.DeleteBlobAsync(productImage.ImageId, "products");
             }
 
+            _context.Products.Remove(product);
+            await _context.SaveChangesAsync();
+            _flashMessage.Info("Registro borrado.");
             return RedirectToAction(nameof(Index));
         }
+
 
     }
 }
